+++ library/src/main/java/com/getstream/sdk/chat/rest/interfaces/DeleteChannelCallback.java
@@ -1,9 +1,0 @@
-package com.getstream.sdk.chat.rest.interfaces;
-
-import com.getstream.sdk.chat.rest.response.DeleteChannelResponse;
-
-public interface DeleteChannelCallback {
-    void onSuccess(DeleteChannelResponse response);
-
-    void onError(String errMsg, int errCode);
-}